+++ flytekit/interfaces/data/local/local_file_proxy.py
@@ -29,6 +29,10 @@ class LocalFileProxy(_common_data.DataProxy):
         """
         self._sandbox = sandbox
 
+    @property
+    def sandbox(self) -> str:
+        return self._sandbox
+
     def exists(self, path):
         """
         :param Text path: the path of the file
