+++ tests/flytekit/unit/annotated/test_flyte_typing.py
@@ -4,6 +4,7 @@ from flytekit.types.flyte_file import FlyteFile
 def test_filepath_equality():
     a = FlyteFile("/tmp")
     b = FlyteFile("/tmp")
+    assert str(b) == "/tmp"
     assert a == b
 
     a = FlyteFile("/tmp")
